requirements.txt
# Macro Liquidity Daily Monitor - Dependencies
# Core (likely pre-installed)
pandas>=1.5
requests>=2.28
matplotlib>=3.6
numpy>=1.23

# Optional (for better YAML parsing)
pyyaml>=6.0
